Revert "Split log4cplus::Initializer into its own header."
This reverts commit c1b80d71421438d920d5581d79894a47a8cdc9e5.
