+++ include/Makefile.in
@@ -375,7 +375,6 @@ nobase_log4cplusinc_HEADERS = \
 	log4cplus/helpers/timehelper.h \
 	log4cplus/hierarchy.h \
 	log4cplus/hierarchylocker.h \
-	log4cplus/initializer.h \
 	log4cplus/internal/cygwin-win32.h \
 	log4cplus/internal/env.h \
 	log4cplus/internal/internal.h \
